requirements.txt
python-dotenv==0.8.2
whatapi==0.2.0